Implement analyzer for platform intrinsics use in System.Private.CoreLib
This analyzer detects the use of all platform intrinsics and checks to
ensure that they are all used either protected by an if statement OR
ternary operator which checks an appropriate IsSupported flag, or that
the intrinsic is used within a method where the behavior of platform
support for the intrinsic is not allowed to vary between compile time
and runtime.

The supported conditional checks are

1. Simple if statement checking IsSupported flag surrounding usage
```
if (PlatformIntrinsicType.IsSupported)
{
    PlatformIntrinsicType.IntrinsicMethod();
}
```

2. If statement check checking a platform intrinsic type which implies
that the intrinsic used is supported.

```
if (Avx2.X64.IsSupported)
{
    Avx2.IntrinsicMethod();
}
```

3. Nested if statement where there is an outer condition which is an
OR'd together series of IsSupported checks for mutually exclusive
conditions and where the inner check is an else clause where some checks
are excluded from applying.

```
if (Avx2.IsSupported || ArmBase.IsSupported)
{
    if (Avx2.IsSupported)
    {
        // Do something
    }
    else
    {
        ArmBase.IntrinsicMethod();
    }
}
```

In addtion, this change adds a new attribute, currently named
System.Runtime.BypassReadyToRunForIntrinsicsHelperUse which can be used
to identify methods which are called expecting to use some particular
set of platform intrinsics. These functions do not need to have explicit
checks, and while this is not yet implemented in the change, the
attribute will change the behavior of crossgen2 so that code will only
be generated if the platform intrinsic support is well known to be
either enabled or disabled at runtime. In addition, the method will only
be compiled or be inlineable if the platform intrinsic support is
enabled that matches at least one of the attributes applied to that
method.