Enable regex to use IndexOf(..., OrdinalIgnoreCase) for prefix searching
As one of its many ways of finding the next possible match starting location, Regex recognizes a string known to start the expression and uses IndexOf to find it.  With this change, it can also do so for OrdinalIgnoreCase.  With improvements to IndexOf(..., OrdinalIgnoreCase), this now yields significantly faster searches through longer inputs, in addition to leading to simpler code in source generated regexes.