Delete support for IAT indirecting base type/interfaces (#85376)
* Delete support for IAT indirecting base type/interfaces

This is used if we have a SharedLibrary.dll a la .NET Native. SharedLibrary.dll model had a lot of problems around versioning and a "framework-dependent" deployment model like that is unlikely to come back. I think we can start to slowly get rid of and get marginally faster code and less maintenance cost in return.

* Bump header