Eliminate cancellation deadlock in RateLimiter implementations
In System.Threading.RateLimiting.ConcurrencyLimiter and the other three
implementations using a synchronously locked deque, there is a deadlock
between an outstanding wait-for-lease operation being canceled by user
and releasing an existing lease. The internal Release() method has to
clean up wait-for-lease operations in some circumstances. This involves
disposing the operation's CancellationTokenRegistration, which blocks
if the registered callback is running. This callback can be invoked on
another thread if the external cancellation token fires. It locks the
rate limiter to return the operation's permits to the rate limiter.
If the external cancellation token fires after Release() runs
but before it disposes of the CancellationTokenRegistration, there will
be a deadlock as Release() holds the lock while waiting for the callback
to complete while the callback blocks on the lock to return its permits.

This change eliminates this deadlock by moving the cleanup of any
wait-for-lease operations in Release() outside the lock.