Rework JIT<->EE communication for struct promotion
Rework the communication so that the JIT asks the EE for struct field
related information relevant to promotion in a single JIT-EE call of a
new function called 'flattenType'.

As a side effect, removes CORINFO_FLG_CUSTOMLAYOUT and
CORINFO_FLG_INDEXABLE_FIELDS (the latter by adding support to inline
arrays in this new function). Also, last but not least, this adds
support for general recursive struct promotion provided the field limit
of 4 is not hit (for now, let's see if there are any more fundamental
issues with this).

Fix #85511
Fix #71711