Reduce overheads of Regex.Replace with a text replacement string (#85564)
* Reduce overheads of Regex.Replace with a text replacement string

If the replacement string doesn't contain any backreferences, we can reduce the overheads involved in processing the replacement.  Rather than storing a list of `ReadOnlyMemory<char>` segments for every portion of the original string or replacement, we can just store a list of `(int offset, int count)` pairs; if the offset is non-negative it refers to the original string, and if it's negative, it means to use the whole replacement.  We can also avoid evaluating the rules each time, and since we're not storing string references into the arrays, we don't need to clear the arrays before returning them to the pool.

* Replace builders with more generic StructListBuilder