Revert unintentional change of one other place involving statics
I have noticed that, probably as part of bulk renames, I have
erroneously changed one other place in getFieldInfo to use the
new helper RvaFieldAddress; this is not appropriate as the place
doesn't deal with RVA fields. This was the reason why several
readytorun/tests failed in the PR testing.

Thanks

Tomas