Merged PR 30131: Only clean up elements that have been marshalled in LibraryImportGenerator
Adds a "paramName_lastIndexMarshalled" variable to generated code that tracks the index of the last element in a collection that was marshalled. Then in the cleanup stage, we only clean up to that index. This works for single dimensional arrays of non-blittable elements.

For multidimensional arrays, this will fully initialize higher order arrays to simplify the generated logic. It is possible to avoid this and can be considered in a future PR, if we observe users of multi-dimensional arrays observing performance issues.