Adjust RVA map offsets to account for wasm prefix
MonoImage:raw_data is used as a base when applying the RVA map to map
virtual addresses to physical offsets in the assembly.  With
webcil-in-wasm there's an extra wasm prefix before the webcil payload
starts, so we need to account for this extra data when creating the
mapping.

An alternative is to compute the correct offsets as part of generating
the webcil, but that would entangle the wasm module and the webcil
payload.  The current (somewhat hacky approach) keeps them
logically separate.