Update mariner images to latest with LLVM 16 (#84676)
This pulls in the updates from
https://github.com/dotnet/dotnet-buildtools-prereqs-docker/commit/53fee55b4115616c7029782c9733055db1bc565a
and moves to floating tags for the mariner images, which now
build the product with LLVM 16.

Once this flows to dotnet-optimization, this will require updates
to make the LLVM 16 tools available there. Before the opt data
collected with new LLVM flows into runtime, I plan to fix
https://github.com/dotnet/runtime/issues/85785 so that local
builds won't by default require clang-16.