Add natstepfilter to skip over object allocators
VS 17.6 added support for natstepfilter in project files. Take advantage of it to skip over object allocators that we pretty much never want to step into.

I'm following up with some VS debugger people to see if there's any plans to allow embedding this in the PDB like natvis. Having this in the vcxproj helps the repro project, but doesn't help end user scenarios.