Refactor our linear collection marshalling implementation
Refactor the elements marshalling logic to split out space allocation and source/destination span construction from the elements marshalling logic. Pull out the element marshalling logic to a common base class and separate it out from the actual marshalling strategies.

This refactoring is required to provide customization points to correctly fix #85795.

I've validated that this change is a zero-diff change on all of the generated code for the LibraryImportGenerator integration tests.