Don't disable AutoreleasePoolSupport by default when enabling trimming.
Disabling AutoreleasePoolSupport is a behavioral change that might affect
apps negatively by increasing memory usage, and it won't be noticed until apps
starts misbehaving by using too much memory (or someone happens to profile
them).

As such, it does not make sense to change the default value for
AutoreleasePoolSupport when trimming, because trimming is supposed to not
change behavior (and certainly not increase memory usage).