Move `mono_object_get_virtual_method` to managed
* Adjust 2 other tests to use a different type now that `Animal` is abstract.

* Delete `can_call_interface_method_on_member_struct` it's problematic and doesn't play nice with the value returned.

This does not support generics.  `MethodInfo.GetMethodFromHandle` needs type information in order to work with generics.  And we don't have that information.

The old implementation of `mono_object_get_virtual_method` was kind of goofy in that for classes, the `if (mono_type_get_type(type) == MONO_TYPE_CLASS)` made it so that `method` was always returned for reference types.