[wasm][bench] Limit initial samples
Try to limit initial samples measurement to 1 second.

Currently some of string measurements with hybrid globalization turned on
can take hundreds of seconds. This makes the browser stop the benchmark
and the measurement job times out.

This should limit the initial samples to take cca 1 second or take just
one initial sample if it takes more time.

Also increase the initial samples for string comparison, where it was on
the edge and the times were flipping between cca 5 and 10 seconds.