[mono][interp] Add fast path intrinsics to some RuntimeType calls
This commit introduces a new type of opcode, call fast paths. This instruction is not emitted via normal codegen, because it needs to exactly preceed the call. Instead, it is stored in the call info data for a MINT_CALL. When emitting the final code for the call instruction we emit also the fast path, if there is one. The fast path instruction will inherit the call return offset and call args offset.

The fast path will preceed the call and will behave as an intrinsic. If the fast path is successful it will store into the return offset the result of the call and bump ip so it skips the follwing call. Otherwise, execution continues normally with the actual call.

This feature would make it easy in the future to add fast path intrinsics for various hot methods, that were too complex to fully intrinsify before.