Add one more suppression to class which marks RUC methods
Recent changes modified the behavior of trimmer where it will now report warnings for all marked methods which have RUC. Previously it tried to avoid duplicate warnings by only doing this for base methods in virtual hierarchies. But that had bugs and the behavior was very complex and hard to predict. So now it warns always.

The DbConnectionStringBuilder class implement ICustomTypeDescriptor interface which has several methods marked with RUC - so the trimmer reported 2113 on all of these. But it didn't warn on the implementation. Now it warns on the implementation as well, with 2112.

The suppression has the same reason as previously, so no functional change.