Optimize zero-byte receives on ManagedWebSocket
Today you can perform a zero-byte receive on ManagedWebSocket, but doing so still issues a request to the underlying stream with the receive header buffer.  That in turn can cause the underlying stream to rent and/or pin a buffer.  By special-casing zero-byte reads, we can take advantage of any special-casing in the base stream, and hopefully make it so that when the actual read is performed, the data necessary to satisfy it synchronously is already available.