Split the responsbility of keeping the objects alive
The WrappedObjectHolder will be used to keep the managed object alive when
the COM reference count is positive. And the ManagedObjectWrapper object
will be responsible for cleaning up the unmanaged MOW when the wrapped
object is truly dead.

The design is that as long as there is a postive reference count, the
WrappedObjectHolder should stay alive. Any time the ref count is increased
from the managed side from 0 to 1, the WrappedObjectHolder will be restored
if it was collected. It should not be possible for the MOW to observe a
null WrappedObjectHolder.