JIT: Add support for induced field accesses in physical promotion
Add support for promoting fields based on the fact that assignment
decomposition induces new field accesses. To do so we store all struct
assignments that involve candidates for physical promotion as part of
the initial accounting pass. Then when picking the replacements we do it
as a fixpoint computation, iteratively expanding the set of replacements
based on field accesses induced by decomposition assignment with the
existing set of replacements.

Fix #87371