SPMI: Unify error handling path for replay and asmdiffs
For asmdiffs we would go into the success path always when the base JIT
succeeded. This meant that if the diff JIT missed we would confusingly
print an error message

    ERROR: method xyz is missing a compileResult, cannot do diffing

even though the "base success, diff miss" case is quite common and
expected. This change makes us take the same error path if either of the
JIT results failed, and then moves the JIT1 unique error handling into
an earlier check.