Perf optimization of module scanning for generic cycle detection
As part of investigation of the bug #66079 and implementation of
the fix #71426 I noticed that one aggravating factor is that
Crossgen2 starts analyzing the same assembly multiple times on
its various parallel threads; this busywork additionally makes the
app compete for access to the same metadata, making the initial
analysis even slower.

In accordance with Michal's suggestion from the PR thread #71426
I propose to modify the generic cycle detector to run the initial
analysis single-threaded if the module in question is expected
to be "generics-heavy" using the number of TypeSpec rows in its
ECMA metadata as an indicator of module generic complexity.

I have written a simple managed app scanning all runtime framework
assemblies, ASP.NET assemblies and assemblies used in internal
CoreCLR testing. I found out that the largest number of TypeSpec
rows is in FSharp.Core (3855), followed by Microsoft.CodeAnalysis
(3148). Based on these findings I have set the initial value of
the cutoff for single-threaded analysis to 5000.

Thanks

Tomas