JIT: Support last-use copy omission for commas
Allow last-use copy omission to kick in when the argument is a comma
(typically created by physical promotion).

Also handle this pattern in arg evaluation: avoid creating a new local
for the COMMA(..., LCL_ADDR) created for this pattern. As part of this
rename the "needs temp" terminology to "needs early evaluation" since
these cases do not actually need or create a temp. There were other
cases already where we did not create a temp (specifically handling of
GT_MKREFANY), so the terminology was already a bit misleading.

Fix #87471