Add support for DOTNET_SYSTEM_NET_HOSTALIASES
Allow an environment variable to specify a host aliases file that's used by System.Net.Dns to resolve the provided names to the strings actually used for resolution.  This enables a set of services to be written to refer to each other via simple names that are then substituted via an aliases file external to the app.