JIT: Insert readbacks eagerly
Physical promotion currently inserts readbacks (reading the struct local
back into the field local) as COMMAs right before a local that needs it.
This is not right for QMARKs where it may mean we only end up reading
back the local in one of the branches.

This change makes physical promotion insert readbacks as new statements
before any statement that is going to need it. While we could do this
for QMARKs only, it is done for any statement indiscriminately since it
has two benefits:
1. It allows forward-sub to kick in for the readbacks, which can lead to
   a contained LCL_FLD
2. It stops us from disabling local copy prop by avoiding the creation
   of embedded stores.

The existing logic is still necessary to keep in case the readback was
marked within the same tree.

Fix #87508